+++ src/server.rs
@@ -6,7 +6,7 @@ use std::{
 
 use quinn::{
     crypto::rustls::QuicServerConfig, default_runtime, Connection, ConnectionError, Endpoint,
-    EndpointConfig, VarInt,
+    EndpointConfig, ReadError, ReadToEndError, VarInt,
 };
 use rustls::pki_types::PrivatePkcs8KeyDer;
 use tokio::{runtime::Handle, sync::oneshot};
@@ -212,6 +212,36 @@ pub enum ClientOnServerDisconnectionDetails {
     Timeout,
 }
 
+fn connection_error_to_client_on_server_event<R: NetReceive>(
+    error: ConnectionError,
+) -> Option<ClientOnServerEvent<R>> {
+    match error {
+        ConnectionError::ApplicationClosed(_thingy) => {
+            // TODO: Deserialize the reason from `_thingy` and put it in the event.
+            Some(ClientOnServerEvent::Disconnected(
+                ClientOnServerDisconnectionDetails::None,
+            ))
+        }
+        ConnectionError::ConnectionClosed(_thingy) => {
+            // TODO: Deserialize the reason from `_thingy` and put it in the event.
+            Some(ClientOnServerEvent::Disconnected(
+                ClientOnServerDisconnectionDetails::None,
+            ))
+        }
+        ConnectionError::LocallyClosed => {
+            // Our own side have closed the connection, let's just wrap up as expected.
+            None
+        }
+        ConnectionError::TimedOut => Some(ClientOnServerEvent::Disconnected(
+            ClientOnServerDisconnectionDetails::Timeout,
+        )),
+        error => {
+            // TODO: Handle more errors to pass as events to the user.
+            panic!("{error}");
+        }
+    }
+}
+
 impl<S: NetSend, R: NetReceive> ClientOnServerNetworking<S, R> {
     fn new(
         async_runtime_handle: Handle,
@@ -229,42 +259,36 @@ impl<S: NetSend, R: NetReceive> ClientOnServerNetworking<S, R> {
                         // to get the entire message that we can then provide to the user.
                         let receiving_sender_cloned = receiving_sender.clone();
                         tokio::spawn(async move {
-                            let message_raw = receive_message_raw(&mut stream).await;
-                            let message: R = rmp_serde::decode::from_slice(&message_raw).unwrap();
-                            let event = ClientOnServerEvent::Message(message);
-                            receiving_sender_cloned.send(event).unwrap();
+                            match receive_message_raw(&mut stream).await {
+                                Ok(message_raw) => {
+                                    // Received all the message successfully!
+                                    let message: R =
+                                        rmp_serde::decode::from_slice(&message_raw).unwrap();
+                                    let event = ClientOnServerEvent::Message(message);
+                                    receiving_sender_cloned.send(event).unwrap();
+                                }
+                                Err(ReadToEndError::Read(ReadError::ConnectionLost(error))) => {
+                                    // Oh we lost the connection in the middle of receiving
+                                    // the message.
+                                    let event = connection_error_to_client_on_server_event(error);
+                                    if let Some(event) = event {
+                                        receiving_sender_cloned.send(event).unwrap();
+                                    }
+                                }
+                                Err(error) => {
+                                    // TODO: Handle more errors to pass as events to the user.
+                                    panic!("{error}");
+                                }
+                            }
                         });
                     }
-                    Err(ConnectionError::ApplicationClosed(_thingy)) => {
-                        // TODO: Deserialize the reason from `_thingy` and put it in the event.
-                        let event = ClientOnServerEvent::Disconnected(
-                            ClientOnServerDisconnectionDetails::None,
-                        );
-                        receiving_sender.send(event).unwrap();
-                        return;
-                    }
-                    Err(ConnectionError::ConnectionClosed(_thingy)) => {
-                        // TODO: Deserialize the reason from `_thingy` and put it in the event.
-                        let event = ClientOnServerEvent::Disconnected(
-                            ClientOnServerDisconnectionDetails::None,
-                        );
-                        receiving_sender.send(event).unwrap();
-                        return;
-                    }
-                    Err(ConnectionError::LocallyClosed) => {
-                        // Our own side have closed the connection, let's just wrap up as expected.
-                        return;
-                    }
-                    Err(ConnectionError::TimedOut) => {
-                        let event = ClientOnServerEvent::Disconnected(
-                            ClientOnServerDisconnectionDetails::Timeout,
-                        );
-                        receiving_sender.send(event).unwrap();
-                        return;
-                    }
                     Err(error) => {
-                        // TODO: Handle more errors to pass as events to the user.
-                        panic!("{error}");
+                        // We just lost the connection.
+                        let event = connection_error_to_client_on_server_event(error);
+                        if let Some(event) = event {
+                            receiving_sender.send(event).unwrap();
+                        }
+                        return;
                     }
                 };
             }